requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic-settings
SQLAlchemy
psycopg[binary]
